requirements.txt
streamlit>=1.30.0
yfinance>=0.2.31
plotly>=5.18.0
pandas>=2.1.0
numpy>=1.25.0
requests>=2.31.0
python-dotenv>=1.0.0
fredapi>=0.5.1
ccxt>=4.1.0
ta>=0.11.0
scipy>=1.11.0
scikit-learn>=1.3.0
pytz>=2023.3
statsmodels>=0.14.0
